+++ headers/item.hpp
@@ -16,23 +16,23 @@ inline constexpr int MAX_CARROT = 16;
 inline constexpr int MAX_APPLE  = 1;
 
 // ---- Textures ----
-inline constexpr int         POTATO_FRAME_W             = 19;
-inline constexpr int         POTATO_FRAME_H             = 19;
-inline constexpr int         POTATO_COLS                = 5;
-inline constexpr int         POTATO_ROW_IDLE            = 0;
+inline constexpr float       POTATO_FRAME_W        = 19.0f;
+inline constexpr float       POTATO_FRAME_H        = 19.0f;
+inline constexpr float       POTATO_COLS           = 5.0f;
+inline constexpr float       POTATO_ROW_IDLE       = 0.0f;
 inline constexpr const char *POTATO_TEXTURE        = "assets/potato.png";
 inline constexpr const char *POTATO_STATIC_TEXTURE = "assets/potato_static.png";
 
-inline constexpr int         CARROT_FRAME_W             = 32;
-inline constexpr int         CARROT_FRAME_H             = 32;
-inline constexpr int         CARROT_COLS                = 4;
-inline constexpr int         CARROT_ROW_IDLE            = 0;
+inline constexpr float       CARROT_FRAME_W        = 32.0f;
+inline constexpr float       CARROT_FRAME_H        = 32.0f;
+inline constexpr float       CARROT_COLS           = 4.0f;
+inline constexpr float       CARROT_ROW_IDLE       = 0.0f;
 inline constexpr const char *CARROT_STATIC_TEXTURE = "assets/carrot.png";
 
-inline constexpr int         APPLE_FRAME_W              = 32;
-inline constexpr int         APPLE_FRAME_H              = 32;
-inline constexpr int         APPLE_COLS                 = 4;
-inline constexpr int         APPLE_ROW_IDLE             = 0;
+inline constexpr float       APPLE_FRAME_W         = 32.0f;
+inline constexpr float       APPLE_FRAME_H         = 32.0f;
+inline constexpr float       APPLE_COLS            = 4.0f;
+inline constexpr float       APPLE_ROW_IDLE        = 0.0f;
 inline constexpr const char *APPLE_STATIC_TEXTURE  = "assets/apple.png";
 
 // ---- Structure pour les objets ----
