+++ headers/player.hpp
@@ -5,11 +5,11 @@
 #include "inventory.hpp"
 
 // ---- Textures ----
-inline constexpr int         PLAYER_FRAME_W  = 32;
-inline constexpr int         PLAYER_FRAME_H  = 32;
-inline constexpr int         PLAYER_COLS     = 4;
-inline constexpr int         PLAYER_ROW_IDLE = 0;
-inline constexpr int         PLAYER_ROW_MOVE = 1;
+inline constexpr float       PLAYER_FRAME_W  = 32.0f;
+inline constexpr float       PLAYER_FRAME_H  = 32.0f;
+inline constexpr float       PLAYER_COLS     = 4.0f;
+inline constexpr float       PLAYER_ROW_IDLE = 0.0f;
+inline constexpr float       PLAYER_ROW_MOVE = 1.0f;
 inline constexpr const char *PLAYER_TEXTURE  = "assets/player.png";
 
 // ---- Classe Player ----
